Most of the major redesign
Pretty much everything has been updated, following the media control bar that was done earlier.
This design moves away from the floating window concept, and cleans up the album and song displays. It's also added a sidebar, this will have links to various routes as well as to each playlist.